more file/dir refactoring
