+++ src/cudnn/mod.rs
@@ -3,12 +3,10 @@ mod algorithm_with_workspace;
 mod batch_normalization;
 mod conv;
 mod cuda_cudnn_result;
-mod data_type;
 mod result;
 #[allow(warnings)]
 mod sys;
 mod tensor;
-mod tensor_ops;
 mod workspace;
 
 pub use activation::*;
@@ -16,8 +14,6 @@ pub use algorithm_with_workspace::*;
 pub use batch_normalization::*;
 pub use conv::*;
 pub use cuda_cudnn_result::*;
-pub use data_type::*;
 pub use result::*;
 pub use tensor::*;
-pub use tensor_ops::*;
 pub use workspace::*;
